[desserbee-web] 디저트메이트 버그 수정 (#212)
* fix: 영업일 선택 시 휴무일 사라지도록 수정

* feat: 운영시간 수정 - 초기화 버튼

* fix: 휴무 주기 줄바꿈되도록 수정

* fix: 휴무주기 매주일 때 다른 옵션들 disabled

* fix: 디저트 메이트 카드 클릭으로 수정, 이벤트 버블링 방지

* fix: 로그인 시 deviceId 쿠키 저장 및 요청마다 전송

* fix: 참여하기 버튼 ui 업데이트 되도록 수정

* fix: 디저트메이트, 리뷰 인증 다시 수정 (content type 충돌)

* fix: 모집완료 된 글에는 버튼 disabled

* fix: authConverter에 deviceId 추가

* fix: 수락/거절 상태 변경 ui에 반영되도록 수정

+++ packages/usecase/src/mateService.ts
@@ -150,7 +150,8 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.create({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.create({ data, authorization });
 
     return response;
   }
@@ -160,7 +161,8 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.update({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.update({ data, authorization });
 
     return response;
   }
@@ -170,7 +172,8 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.delete({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.delete({ data, authorization });
 
     return response;
   }
@@ -181,7 +184,6 @@ export default class MateService {
     }
 
     const authorization = await this.authRepository?.getAuthorization();
-
     const response = await this.mateRepository.save({ data, authorization });
 
     return response;
@@ -193,7 +195,6 @@ export default class MateService {
     }
 
     const authorization = await this.authRepository?.getAuthorization();
-
     const response = await this.mateRepository.cancelSave({
       data,
       authorization,
@@ -207,7 +208,11 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.acceptMyTeamMember({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.acceptMyTeamMember({
+      data,
+      authorization,
+    });
 
     return response;
   }
@@ -217,7 +222,11 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.rejectMyTeamMember({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.rejectMyTeamMember({
+      data,
+      authorization,
+    });
 
     return response;
   }
@@ -227,7 +236,8 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.write({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.write({ data, authorization });
 
     return response;
   }
@@ -237,7 +247,8 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.edit({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.edit({ data, authorization });
 
     return response;
   }
@@ -247,7 +258,11 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.createReply({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.createReply({
+      data,
+      authorization,
+    });
 
     return response;
   }
@@ -259,7 +274,11 @@ export default class MateService {
       throw new Error('mateRepository is not set');
     }
 
-    const response = await this.mateRepository.getReplyList({ data });
+    const authorization = await this.authRepository?.getAuthorization();
+    const response = await this.mateRepository.getReplyList({
+      data,
+      authorization,
+    });
 
     return response;
   }
@@ -274,7 +293,6 @@ export default class MateService {
     }
 
     const authorization = await this.authRepository?.getAuthorization();
-
     const response = await this.mateRepository.getSavedMateList({
       data,
       authorization,
